services/orchestrator/requirements.txt
# API et serveur
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
httpx<0.17.0,>=0.16.1

# Base de données
supabase-py==0.0.2
asyncpg==0.29.0

# Utilitaires
python-dotenv==1.0.0
loguru==0.7.2
tenacity==8.2.3
pytz==2023.4

# Tests
pytest<7,>=6
pytest-asyncio<7
pytest-cov==4.1.0

# Dev tools
black==24.1.1
isort==5.13.2
flake8==7.0.0
pyyaml==6.0